Repositorios y Servicio de eventos agreado

+++ src/main/java/com/smarroquin/EventBook/models/Event.java
@@ -8,7 +8,6 @@ public class Event {
     private Date date;
     private String location;
 
-
     public Event(long id, String name, Date date, String location) {
         this.id = id;
         this.name = name;
@@ -16,13 +15,13 @@ public class Event {
         this.location = location;
     }
 
-    public long getId() {return id;}
-    public String getName() {return name;}
-    public Date getDate() {return date;}
-    public String getLocation() {return location;}
+    public long getId() { return id; }
+    public String getName() { return name; }
+    public Date getDate() { return date; }
+    public String getLocation() { return location; }
 
-    public void setId(int id) {this.id = id;}
-    public void setName(String name) {this.name = name;}
-    public void setDate(Date date) {this.date = date;}
-    public void setLocation(String location) {this.location = location;}
+    public void setId(long id) { this.id = id; }
+    public void setName(String name) { this.name = name; }
+    public void setDate(Date date) { this.date = date; }
+    public void setLocation(String location) { this.location = location; }
 }
